Make `waitFor` more reliable
If somebody does two synchronous updates in a row, the `nextPromise` will not re-arm itself for the second update, and the `waitFor` function will only see the first value. Switching from the promise value to the cached value avoids this problem, since `waitFor` will always see the newest props available.